src/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx
openpyxl
pillow
easyocr
pandas
numpy